added orderItem logic incase the product doesn't exist

+++ store/views.py
@@ -38,29 +38,30 @@ def checkout(request):
 	return render(request, 'store/checkout.html', context)
 
 def updateItem(request):
-	data = json.loads(request.body)
-	productId = data['productId']
-	action = data['action']
-	print('Action:', action)
-	print('Product:', productId)
+    data = json.loads(request.body)
+    productId = data['productId']
+    action = data['action']
+    print('Action:', action)
+    print('Product:', productId)
 
-	customer = request.user.customer
-	product = Product.objects.get(id=productId)
-	order, created = Order.objects.get_or_create(customer=customer, complete=False)
+    customer = request.user.customer
+    product = Product.objects.filter(id=productId).first()  # Use .first() to prevent NoneType errors
 
-	orderItem, created = OrderItem.objects.get_or_create(order=order, product=product)
+    if product:  # Only proceed if product exists
+        order, created = Order.objects.get_or_create(customer=customer, complete=False)
+        orderItem, created = OrderItem.objects.get_or_create(order=order, product=product)
 
-	if action == 'add':
-		orderItem.quantity = (orderItem.quantity + 1)
-	elif action == 'remove':
-		orderItem.quantity = (orderItem.quantity - 1)
+        if action == 'add':
+            orderItem.quantity = (orderItem.quantity + 1)
+        elif action == 'remove':
+            orderItem.quantity = (orderItem.quantity - 1)
 
-	orderItem.save()
+        orderItem.save()
 
-	if orderItem.quantity <= 0:
-		orderItem.delete()
+        if orderItem.quantity <= 0:
+            orderItem.delete()
 
-	return JsonResponse('Item was added', safe=False)
+    return JsonResponse('Item was added', safe=False)
 
 def processOrder(request):
 	transaction_id = datetime.datetime.now().timestamp()
